Split mobile views into api/mobile_views.py, fix ResetFace/ConfidentUnlabeledView
Moved the 4 mobile-tagging-app views (ConfidentUnlabeledView,
UnlabeledMobileInfo, ResetFace, MobileNameList) out of api/views.py --
they were the entire /api/mobile/... URL group, mixed in with the
standard ModelViewSets and slideshow-facing endpoints in an already
large file. api/urls.py now imports them from api.mobile_views.

Fixed two bugs while moving:
- ResetFace.patch() had no return statement, so DRF's dispatch() got
  None back instead of a Response and raised AssertionError -- crashed
  on every single call. Now returns a small JSON success body. Also
  dropped a leftover @action(...) decorator, which is a DRF-router-only
  decorator that does nothing on a plain APIView.
- ConfidentUnlabeledView.get() did unlabeled[0].weight_1/unlabeled.last()
  .weight_1 unconditionally (an unused sanity-check assert, not part of
  the response) -- unlabeled[0] raised IndexError the moment there were
  zero unlabeled faces, the goal state of the tagging workflow, not a
  rare edge case. Now just returns whatever ids exist, including none.

MobileNameList (still a hardcoded-stub, separate known issue) and
UnlabeledMobileInfo moved as-is, untouched otherwise.

Not yet ported to master/deployed -- tracked as a TODO in CLAUDE.md's
Planned work.
